--- a/lib/web_ui/dev/roll_fallback_fonts.dart
+++ b/lib/web_ui/dev/roll_fallback_fonts.dart
@@ -175,7 +175,6 @@ class RollFallbackFontsCommand extends Command<bool>
             'Unexpected url format received from Google Fonts API: $urlString.');
       }
       final String urlSuffix = urlString.substring(expectedUrlPrefix.length);
-      sb.writeln(" // ${font.shortName}: ${charsetForFamily[family]}");
       sb.writeln(" NotoFont('$family', $enabledArgument'$urlSuffix'),");
     }
     sb.writeln('];');