--- a/lib/web_ui/lib/src/engine/font_fallback_data.dart
+++ b/lib/web_ui/lib/src/engine/font_fallback_data.dart
@@ -23418,4 +23418,3 @@ const String encodedFontSetRanges =
     'iP'       // fe82e-fe837  #15
     '2754sE'   // fe838-10ffff
     ;
-